회원가입 api 구현. 레디스 의존성 제거, Header.java 삭제 --> ResponseEntity로 대체, 프론트엔드 디렉터리 삭제
