+++ build.gradle
@@ -39,8 +39,6 @@ dependencies {
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.security:spring-security-test'
-
-    runtimeOnly 'com.h2database:h2' // h2 DB
 }
 
 def webappDir = "$projectDir/src/main/frontend/studymaker"
