--- src/main/frontend/studymaker/src/App.css
@@ -1,226 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: "Noto Sans KR", sans-serif;
-  letter-spacing: -0.75px;
-}
-
-ul {
-  list-style: none;
-}
-
-a {
-  text-decoration: none;
-  outline: none;
-}
-
-a:link,
-a:visited {
-  color: #767676;
-}
-
-a:hover,
-a:active {
-  color: #3c91e6;
-}
-
-label {
-  font-size: 15px;
-}
-
-/* signUp page */
-.signup-container {
-  display: grid;
-  grid-gap: 2.5rem;
-  place-items: center;
-  margin-top: 30px;
-}
-
-.container > h2 {
-  margin-right: 210px;
-}
-
-header {
-  width: 100%;
-  height: 65px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 6rem;
-  box-shadow: 0 2px 5px 0 #dbdbdb;
-}
-
-#logo {
-  color: #3c91e6;
-  font-size: 18px;
-}
-
-header > nav {
-  width: 180px;
-  height: 100%;
-}
-
-header ul {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: space-between;
-}
-
-header ul > li {
-  height: 100%;
-  display: flex;
-  align-items: center;
-}
-
-header ul > li > a {
-  font-size: 1rem;
-}
-
-.form {
-  display: grid;
-  grid-row-gap: 0.5rem;
-}
-
-.form > div {
-  display: grid;
-  grid-gap: 5px;
-}
-
-.input {
-  border: 0.7px solid #ededed;
-  border-radius: 5px;
-  padding: 10px 80px 10px 8px;
-  color: #000;
-}
-
-.error-message {
-  color: red;
-  font-size: 13px;
-}
-
-.no-error {
-  opacity: 0;
-  font-size: 13px;
-}
-
-.double-check {
-  border: 0.7px solid #3c91e6;
-  border-radius: 3px;
-  margin-left: 5px;
-  background: none;
-  font-size: 12px;
-  padding: 10px 5px;
-  color: #3c91e6;
-  cursor: pointer;
-}
-
-#button-div {
-  display: inline-block;
-  text-align: center;
-}
-
-#button-div button {
-  border: 0.7px solid #3c91e6;
-  border-radius: 3px;
-  font-size: 15px;
-  padding: 10px 40px;
-  cursor: pointer;
-}
-
-#back-button {
-  background: none;
-  margin-right: 8px;
-  color: #3c91e6;
-}
-
-#signup-button {
-  background: #3c91e6;
-  margin-left: 8px;
-  color: #fff;
-  box-shadow: 0 3px 7px 0 #dbdbdb;
-}
-
-/* main page */
-.main-container {
-  display: grid;
-  grid-gap: 2rem;
-  place-items: center;
-  margin-top: 30px;
-}
-
-.main-category {
-  display: grid;
-  grid-gap: 1.5rem;
-}
-
-.category-title {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.category-title > p > a {
-  font-size: 14px;
-}
-
-.studyCardList {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-gap: 1.5rem;
-}
-
-.studyCard {
-  border: 1px solid #ededed;
-  border-radius: 5px;
-}
-
-.study-thumbnail {
-  border-radius: 5px 5px 0 0;
-}
-
-.study-detail {
-  padding: 5px 10px 20px 10px;
-}
-
-.study-name {
-  font-size: 15px;
-  color: #3c91e6;
-  font-weight: bold;
-}
-
-.study-summary {
-  font-size: 12px;
-  color: #767676;
-  margin-bottom: 20px;
-}
-
-.study-env {
-  font-size: 11px;
-  color: #767676;
-}
-
-footer {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  height: 80px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  padding: 0 6rem;
-}
-
-footer ul > li {
-  display: inline-block;
-  padding: 0 5px;
-}
-
-footer ul > li > a {
-  font-size: 0.8rem;
-}
