--- src/main/frontend/studymaker/src/app.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-import "./app.css";
-import Main from "./components/main";
-import SignUp from "./components/signUp";
-
-function App() {
-
-  
-  return (
-    <>
-      <Main />
-    </>
-  );
-}
-
-export default App;
