--- src/main/frontend/studymaker/src/components/footer.jsx
@@ -1,32 +1,0 @@
-import React, { Component } from "react";
-
-class Footer extends Component {
-  render() {
-    return (
-      <footer>
-        <ul>
-          <li>
-            <a href="">이용약관</a>
-          </li>
-          <li>
-            <a href="">개인정보처리방침</a>
-          </li>
-          <li>
-            <a href="">커뮤니티이용규칙</a>
-          </li>
-          <li>
-            <a href="">문의하기</a>
-          </li>
-          <li>
-            <a href="">공지사항</a>
-          </li>
-          <li>
-            <a href="#">©스터디메이커</a>
-          </li>
-        </ul>
-      </footer>
-    );
-  }
-}
-
-export default Footer;
