--- src/main/frontend/studymaker/src/components/studyCardList.jsx
@@ -1,15 +1,0 @@
-import React from "react";
-import StudyCard from "./studyCard";
-
-const StudyCardList = () => {
-  return (
-    <div className="studyCardList">
-      <StudyCard />
-      <StudyCard />
-      <StudyCard />
-      <StudyCard />
-    </div>
-  );
-};
-
-export default StudyCardList;
