--- src/main/java/com/anytime/studymaker/Controller/UserController.java
@@ -1,27 +1,0 @@
-package com.anytime.studymaker.Controller;
-
-import com.anytime.studymaker.domain.common.Header;
-import com.anytime.studymaker.domain.user.dto.UserApiRequest;
-import com.anytime.studymaker.domain.user.dto.UserApiResponse;
-import com.anytime.studymaker.service.user.UserService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RequiredArgsConstructor
-@RequestMapping("/api/user")
-@RestController
-public class UserController {
-
-    private final UserService userService;
-
-    @PostMapping("/signUp")
-    public Header<UserApiResponse> signUp(UserApiRequest request) {
-        //    TODO : 회원가입
-
-        return new Header<UserApiResponse>();
-    }
-}
