--- src/main/java/com/anytime/studymaker/Controller/common/CrudController.java
@@ -1,4 +1,0 @@
-package com.anytime.studymaker.Controller.common;
-
-public interface CrudController {
-}
