--- src/main/java/com/anytime/studymaker/config/data/RedisConfig.java
@@ -1,39 +1,0 @@
-package com.anytime.studymaker.config.data;
-
-import com.anytime.studymaker.config.jwt.Token;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-@EnableRedisRepositories(basePackages = "com.anytime.studymaker.domain.**.repository.redis")
-public class RedisConfig {
-
-    @Value("${spring.redis.host}")
-    private String redisHost;
-
-    @Value("${spring.redis.port}")
-    private int redisPort;
-
-    @Bean
-    public RedisConnectionFactory redisConnectionFactory() {
-        return new LettuceConnectionFactory(redisHost, redisPort);
-    }
-
-    @Bean
-    public RedisTemplate<String, Object> redisTemplate() {
-        RedisTemplate<String, Object> template = new RedisTemplate<>();
-        template.setConnectionFactory(redisConnectionFactory());
-        template.setKeySerializer(new StringRedisSerializer());
-
-//        객체를 JSON 형태로 받기 위한 직렬화 작업
-        template.setValueSerializer(new Jackson2JsonRedisSerializer<>(Token.class));
-        return template;
-    }
-}
