+++ src/main/java/com/anytime/studymaker/controller/UserController.java
@@ -1,0 +1,14 @@
+package com.anytime.studymaker.controller;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RequiredArgsConstructor
+@RequestMapping("/api/user")
+@RestController
+public class UserController {
+
+}
