+++ src/main/java/com/anytime/studymaker/controller/common/CrudController.java
@@ -1,0 +1,4 @@
+package com.anytime.studymaker.controller.common;
+
+public interface CrudController {
+}
