+++ src/main/java/com/anytime/studymaker/domain/user/User.java
@@ -8,7 +8,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
-import org.springframework.data.redis.core.RedisHash;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
@@ -24,10 +23,8 @@ import java.util.*;
 @Builder
 @Data
 @Accessors(chain = true)
-@RedisHash("studymaker")
 @Entity
 public class User implements UserDetails {
-    @org.springframework.data.annotation.Id
     @Id
     @GeneratedValue
     private Long userId;
